ENH make check_test fixture atomated in the benchmark's tests
